Assignment2/Resume_Builder_monolithic/requirements.txt
Flask==2.0.1
google-generativeai==0.7.2
python-dotenv==0.19.0
Werkzeug==2.2.2
PyPDF2==3.0.1
python-docx==0.8.11
pytest==6.2.5
reportlab==4.2.2